ability to insert/delete table column was added to table model + couple of tests created.

+++ core/languageDesign/editor/source/jetbrains/mps/nodeEditor/cells/EditorCell_Label.java
@@ -621,8 +621,10 @@ public abstract class EditorCell_Label extends EditorCell_Basic {
   private void deleteIfPossible() {
     if ("".equals(getText()) && isTopCell()) {
       if (getStyle().get(StyleAttributes.AUTO_DELETABLE)) {
-        getSNode().delete();
-      } 
+        if (!executeAction(CellActionType.DELETE)) {
+          getSNode().delete();
+        }
+      }
     }
   }
